Step4 완료했습니다~ (#3722)
* feat : null 또는 빈 문자열이라면 예외 기능 추가

* feat : 덧셈 기능 추가

* feat : 뺄셈 기능 추가

* feat : 곱셈, 나눗셈, 복잡한 사칙연산 기능 추가

* feat : 정의된 연산자 이외라면 예외 기능 추가 및 연산자 enum 분리

* refactor : Calculate 클래스 제거 및 사칙연산 과정을 함수형 인터페이스를 활용하여 단순화

* feat : 연산자가 양끝에 있는 경우 예외 발생하는 기능 추가

* style : 코드 포멧팅 적용

* refactor : toOperator stream 써서 리팩토링

* refactor : FunctionalInterface를 직접 구현하도록 수정

* refactor : CalculateHandler 분리

* chore : 위치 변경

* feat :  Cash 추가

* feat : 두 개의 배열을 비교하여 일치 개수를 반환하는 Match 추가

* feat : 누적 수익금과 수익률을 계산해주는 Profit 추가

* feat : Cash에 toString 재정의

* feat : MatchCache에 toString 정의

* feat : Profit에 toString 정의

* chore : 테스트 코드들 위치 변경

* style : 코드 포멧팅 적용

* refactor : 출력, 입력 관련된 로직들을 InputView, OutputView로 이동 후 toString 삭제

* feat : 보너스 당첨 여부에 따른 금액 변동 기능 추가

* docs : 리팩토링 목록 작성

* refactor : 원시값 bonusNumber와 컬렉션 winningNumbers를 합친 WinningNumber 클래스 추가

* feat : WinningNumber 유효성 처리 추가

* style : 코드 포멧팅

* refactor : WinningLotto 리팩토링
이름 변경
getter를 쓰지 않고 메시지를 보내는 형태로 리팩토링

* docs : step4 Todo.md 추가

* feat : 수동으로 구매할 로또 개수가 null이면 예외 기능 추가

* feat : 수동으로 구매 가능한 제한을 넘기면 예외 기능 추가

* feat : 수동으로 구매하는 번호들 중복 검사 기능 추가

* feat : 수동으로 구매하는 번호들이 범위 체크 기능 추가

* refactor : ManualLotto 내의 원시값을 ManualLottoNumber로 수정

* feat : 수동으로 구매하는 번호 개수 체크 기능 추가

* feat : 수동 구매 로또 개수 입출력 추가

* feat : 수동 로또 번호들 입력 기능 추가

* feat : ManualLottoGroup을 Lotto 변환 작업

* feat : 수동 개수, 자동 개수 출력 기능 추가

* chore : 코드 병합 과정에서 충돌난 것 해결

* style : 코드 포맷팅 적용

* docs : 코드 리뷰 Todo 작성

* chore : Lotto 이름 변경(Lottos)

* chore : LottoNumber 이름 변경(Lotto)

* chore : ManualLottoNumber 이름 변경(LottoNumber)

* refactor : 기존 Manual 클래스들 모두 기존 Lotto와 병합, LottoGroup 삭제

* refactor : LottoNumber 객체 추가에 따른 WinningLotto에 변경사항들 반영

* refactor : 수익률 데이터 타입 변경

* refactor : printPlainMessage 삭제

* refactor : Cash 내부의 minus에 예외 처리 추가

* docs : 코드리뷰 반영 todo 작성

* fix : Index Out of Range 오류 해결

* fix : 수익률 계산 오류 수정

* refactor : 모든 예외에 메시지 추가

* refactor : Lotto 생성자 내의 null 체킹 메서드 분리

* refactor : LottoClient에 있던 출력관련 로직들을 모두 ViewUtil에 분리

+++ src/main/java/lotto/domain/Profit.java
@@ -3,7 +3,7 @@ package lotto.domain;
 import java.util.Objects;
 
 public class Profit {
-    private int winningPrize;
+    private long winningPrize;
 
     public Profit(int winningPrize) {
         this.winningPrize = winningPrize;
